+++ billinghelper/src/main/kotlin/com/vojtkovszky/billinghelper/PriceUtil.kt
@@ -2,94 +2,94 @@
 
 package com.vojtkovszky.billinghelper
 
+import android.annotation.SuppressLint
 import android.util.Log
 import com.android.billingclient.api.ProductDetails
-import java.util.regex.Pattern
-import kotlin.math.roundToLong
 
-object PriceUtil {
-    internal const val TAG = "BillingHelper/PriceUtil"
-
+internal object PriceUtil {
     var enableLogging: Boolean = false
-
-    /**
-     * Convert a formatted price to a formatted price with a given divider.
-     * For example: "12.80 EUR" with a divider of 4 will return "4.20 EUR"
-     *
-     * @param price formatted price input
-     * @param divider divide the extracted price
-     * @return formatted divided price or null on error
-     */
-    fun convertFullPriceToDivided(
-        price: String,
-        divider: Int
-    ): String? {
-        var fullPrice = price
-        if (divider == 1) {
-            return price
-        } else try {
-            // Must use javaSpaceChar and not a typed space
-            fullPrice = fullPrice.replace("(?<=\\d)\\p{javaSpaceChar}+(?=\\d)".toRegex(), "").trim()
-            fullPrice =
-                if (fullPrice.contains(",") && fullPrice.contains(".") && fullPrice.last() != '.')
-                    fullPrice.replace(",", "")
-                else fullPrice.replace(",", ".")
-
-            var digit: String? = null
-            val currency: String
-            val regex = Pattern.compile("(\\d+(?:\\.\\d+)?)")
-            val matcher = regex.matcher(fullPrice)
-            while (matcher.find()) {
-                digit = matcher.group(1)
-            }
-            if (digit != null) {
-                currency = fullPrice.replace(digit, "")
-                val digitValue = digit.toDouble() / divider
-
-                return if (fullPrice.startsWith(currency)) currency + roundDigitString(digitValue)
-                else roundDigitString(digitValue) + currency
-            }
-        } catch (e: Exception) {
-            if (enableLogging) Log.e(TAG, e.message ?: "")
-        }
-
-        return null
-    }
-
-    private fun roundDigitString(digitValue: Double): String {
-        val priceValueString =
-            if (digitValue > 1000.0) digitValue.roundToLong().toString() else
-                String.format("%.2f", digitValue)
-        return priceValueString.replace(".", ",")
-    }
+    const val TAG = "BillingHelper/PriceUtil"
 }
-
 
 /**
  * A helper method making it easier to retrieve a formatted price from product details.
  * Parameters only apply to subscription products.
  *
- * @param subscriptionOfferIndex index of [ProductDetails.getSubscriptionOfferDetails].
- * @param subscriptionPricingPhaseIndex index of [ProductDetails.SubscriptionOfferDetails.pricingPhases]
+ * @param subscriptionBasePlanId applies to a subscription, define base plan id to get price for. If no value is applied, first offer will be used. Can be combined with [subscriptionOfferId]
+ * @param subscriptionOfferId applies to a subscription, define offer id to get price for. If no value is applied, first offer will be used. Can be combined with [subscriptionBasePlanId]
+ * @param subscriptionPricingPhaseIndex in case your offer has multiple phases, retrieve price for a phase with a given index.
  * @return formatted price or null on error
  */
 fun ProductDetails.getFormattedPrice(
-    subscriptionOfferIndex: Int = 0,
+    subscriptionBasePlanId: String? = null,
+    subscriptionOfferId: String? = null,
     subscriptionPricingPhaseIndex: Int = 0
 ): String? {
     return if (isInAppPurchase()) {
         oneTimePurchaseOfferDetails?.formattedPrice
     } else try {
         if (isSubscription()) {
-            subscriptionOfferDetails?.getOrNull(subscriptionOfferIndex)
-                ?.pricingPhases
-                ?.pricingPhaseList?.getOrNull(subscriptionPricingPhaseIndex)
-                ?.formattedPrice
+            // Find the first matching offer based on the base plan ID and offer ID, or default to the first
+            val offer = subscriptionOfferDetails?.firstOrNull { offer ->
+                (subscriptionBasePlanId == null || offer.basePlanId == subscriptionBasePlanId) &&
+                        (subscriptionOfferId == null || offer.offerId == subscriptionOfferId)
+            }
+            return offer?.pricingPhases?.pricingPhaseList?.getOrNull(subscriptionPricingPhaseIndex)?.formattedPrice
         } else {
             null
         }
     } catch (e: Exception) {
         if (PriceUtil.enableLogging) Log.e(PriceUtil.TAG, e.message ?: "")
         null
+    }
+}
+
+/**
+ * Same as [getFormattedPrice], but apply divider to the actual price.
+ * For example: "16.80 EUR" with a divider of 4 will return "4.20 EUR".
+ * Formatted price will be rounded to two decimal places.
+ *
+ * @param subscriptionBasePlanId see [getFormattedPrice]
+ * @param subscriptionOfferId see [getFormattedPrice]
+ * @param subscriptionPricingPhaseIndex see [getFormattedPrice]
+ * @param divider price divider
+ * @param dividerFormat divider format used for divided price when represented as String.
+ *                      Defaults to two decimal places (as "%.2f") but use other format if needed.
+ */
+@SuppressLint("DefaultLocale")
+fun ProductDetails.getFormattedPriceDivided(
+    subscriptionBasePlanId: String? = null,
+    subscriptionOfferId: String? = null,
+    subscriptionPricingPhaseIndex: Int = 0,
+    divider: Int,
+    dividerFormat: String = "%.2f"
+): String? {
+    return getFormattedPrice(
+        subscriptionBasePlanId,
+        subscriptionOfferId,
+        subscriptionPricingPhaseIndex
+    )?.let { price ->
+        // Regex to match currency before or after the numeric part, with comma or dot as decimal separator
+        val regex = """(\D*)\s*([\d,.]+)\s*(\D*)""".toRegex()
+        val matchResult = regex.matchEntire(price)
+
+        return if (matchResult != null) {
+            val (currencyPrefix, numericPart, currencySuffix) = matchResult.destructured
+
+            // Clean numeric part by replacing comma with dot if needed and converting to Double
+            val normalizedNumber = numericPart.replace(",", ".").toDoubleOrNull()
+            if (normalizedNumber != null) {
+                // Divide and format the result with two decimal places
+                val dividedPrice = normalizedNumber / divider
+                val formattedPrice = String.format(dividerFormat, dividedPrice)
+
+                // Construct the new price string with currency in the original position
+                "${currencyPrefix.trim()}$formattedPrice${currencySuffix.trim()}"
+            } else {
+                null // Return null if numeric conversion fails
+            }
+        } else {
+            null // Return null if the price format is not as expected
+        }
     }
 }