+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 3.3.0 (2024-11-11)
+* Added optional `subscriptionPurchaseParams` to `launchPurchaseFlow` method to unify subscription purchase definitions. 
+  Added optional parameters `basePlanId: String?` and `offerId: String?`, which replaced old index parameters. 
+  Similarly with `ProductDetails.getFormattedPrice`, which now take optional `subscriptionBasePlanId` and `subscriptionOfferId`.
+* Removed `PriceUtil` object, use extension methods `ProductDetails.getFormattedPrice` and `ProductDetails.getFormattedPriceDivided`.
+
 ## 3.2.0 (2024-11-05)
 * Support subscription updates or replacements: Added optional parameters `subscriptionUpdateOldToken`, 
   `subscriptionUpdateExternalTransactionId` and `subscriptionUpdateReplacementMode` to `launchPurchaseFlow()`